+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Unreleased section should follow [Release Toolkit](https://github.com/newrelic/r
 
 ### Bugfix
 - Fetch the `SoftLimitBytes` metric from docker inspect for cgroups-v2.
+- Fetch the `CPUShares` metric from docker inspect for cgroups-v2.
 
 ## v2.0.7 - 2024-08-06